Move the env plumbing code to short-lived-token.ts
